# src/main/resources/openapi/api.yaml
openapi: 3.0.3
info:
  title: Product API
  description: |
    A RESTful API for managing products using a contract-first approach.
    This API demonstrates OpenAPI specification, Flyway migrations, and PostgreSQL integration.
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com

servers:
  - url: http://localhost:8080/api/v1
    description: Local development server

tags:
  - name: Products
    description: Product management operations

paths:
  /products:
    get:
      tags:
        - Products
      summary: Get all products
      description: Retrieves a list of all products in the system
      operationId: getAllProducts
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    post:
      tags:
        - Products
      summary: Create a new product
      description: Creates a new product with the provided information
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  
  /products/{id}:
    get:
      tags:
        - Products
      summary: Get product by ID
      description: Retrieves a specific product by its ID
      operationId: getProductById
      parameters:
        - name: id
          in: path
          required: true
          description: Product ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    put:
      tags:
        - Products
      summary: Update product
      description: Updates an existing product with the provided information
      operationId: updateProduct
      parameters:
        - name: id
          in: path
          required: true
          description: Product ID
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
      responses:
        '200':
          description: Product updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    delete:
      tags:
        - Products
      summary: Delete product
      description: Deletes a product by its ID
      operationId: deleteProduct
      parameters:
        - name: id
          in: path
          required: true
          description: Product ID
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Product deleted successfully
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    Product:
      type: object
      required:
        - id
        - name
        - price
        - quantity
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier for the product
          example: 1
        name:
          type: string
          description: Product name
          minLength: 1
          maxLength: 100
          example: "Laptop"
        description:
          type: string
          description: Product description
          maxLength: 500
          example: "High-performance laptop with 16GB RAM"
        price:
          type: number
          format: double
          description: Product price
          minimum: 0
          example: 999.99
        quantity:
          type: integer
          format: int32
          description: Available quantity in stock
          minimum: 0
          example: 50
    
    ProductRequest:
      type: object
      required:
        - name
        - price
        - quantity
      properties:
        name:
          type: string
          description: Product name
          minLength: 1
          maxLength: 100
          example: "Laptop"
        description:
          type: string
          description: Product description
          maxLength: 500
          example: "High-performance laptop with 16GB RAM"
        price:
          type: number
          format: double
          description: Product price
          minimum: 0
          example: 999.99
        quantity:
          type: integer
          format: int32
          description: Available quantity in stock
          minimum: 0
          example: 50
    
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
        status:
          type: integer
          description: HTTP status code
        error:
          type: string
          description: Error type
        message:
          type: string
          description: Error message
        path:
          type: string
          description: Request path